test(wasm): expand test suite (48 → 85) and add Firefox CI matrix
Restructure and extend the wasm-bindgen test suite per the approved
test-suite expansion plan:

Structural
- Split existing tests from searchlite-wasm/src/wasm.rs into focused
  integration files under searchlite-wasm/tests/ grouped by surface:
  lifecycle, ingest_delete_update, query, maintenance, migration,
  worker, errors. Shared fixtures now live in tests/common/mod.rs.
- Keep 11 tests inline in src/wasm.rs that require private-state
  access: JsStorage internals, schema_hash, PersistQuotaFailureGuard,
  MigrationFailureGuard, and the internal core-parity comparison.
- src/wasm.rs drops from 4,185 lines to 2,907 lines.

New coverage (+37 tests)
- lifecycle: list_indexes happy-path, drop_index/clear_index reserved
  name guards, drop_index idempotent for missing db, cleanup_indexes
  negative/NaN duration, cleanup_indexes dry-run preserves state.
- ingest_delete_update: add_document non-object, add_documents scalar,
  add_document non-serializable value (JsValue::UNDEFINED),
  update_document non-object request, delete_document with unknown id.
- query: search_controlled / search_request_controlled happy paths,
  search_request malformed/wrong-shape rejection, async variant
  aborted/timeout propagation.
- maintenance: stats reflects ingest/delete counts, inspect redacts
  write-key metadata, cleanup_orphaned_files on clean index, dry_run
  flag preservation.
- migration: plan_migration for missing index is read-only (regression
  for d916710), plan/migrate invalid_schema_json, migrate_index
  preserves registry on compatible schema.
- worker: SearchliteWorkerClient end-to-end add+search roundtrip,
  resetIndex clears content, flushStorage idempotent, storageUsage
  roundtrip, dispose rejects pending calls.
- errors: dedicated per-error-type coverage for invalid_json,
  invalid_schema_json, invalid_update_request, invalid_mget_request,
  invalid_argument.

CI
- Add test-wasm-browser-firefox job running
  wasm-pack test --headless --firefox searchlite-wasm. geckodriver
  ships out of the box on ubuntu-latest. The existing Chrome job is
  unchanged.
