+++ .github/workflows/ci.yml
@@ -56,6 +56,26 @@ jobs:
           RUST_TEST_THREADS: "1"
         run: wasm-pack test --headless --chrome searchlite-wasm
 
+  test-wasm-browser-firefox:
+    name: WASM Browser (Firefox)
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: dtolnay/rust-toolchain@stable
+        with:
+          targets: wasm32-unknown-unknown
+      - uses: Swatinem/rust-cache@v2
+        with:
+          cache-on-failure: true
+      - name: Install wasm-pack
+        run: cargo install wasm-pack --locked
+      - name: Build WASM package
+        run: wasm-pack build searchlite-wasm --target web --release
+      - name: Run browser tests
+        env:
+          RUST_TEST_THREADS: "1"
+        run: wasm-pack test --headless --firefox searchlite-wasm
+
   test-node:
     name: Node.js
     runs-on: ubuntu-latest
