+++ searchlite-wasm/src/wasm.rs
@@ -2442,7 +2442,6 @@ impl Searchlite {
 mod tests {
   use super::*;
   use searchlite_core::api::types::ExecutionStrategy;
-  use std::cell::Cell;
   use std::io::{Read, Seek, SeekFrom, Write};
   use wasm_bindgen_test::*;
 
@@ -2450,195 +2449,6 @@ mod tests {
 
   fn unique_db(name: &str) -> String {
     format!("{name}-{}", js_sys::Date::now() as u64)
-  }
-
-  fn set_timeout_once(ms: i32, callback: impl FnOnce() + 'static) {
-    let cb = Closure::once(callback);
-    web_sys::window()
-      .unwrap()
-      .set_timeout_with_callback_and_timeout_and_arguments_0(cb.as_ref().unchecked_ref(), ms)
-      .unwrap();
-    cb.forget();
-  }
-
-  async fn fetch_ok(path: &str) -> bool {
-    let Some(window) = web_sys::window() else {
-      return false;
-    };
-    let response = match JsFuture::from(window.fetch_with_str(path)).await {
-      Ok(response) => response,
-      Err(_) => return false,
-    };
-    js_sys::Reflect::get(&response, &JsValue::from_str("ok"))
-      .ok()
-      .and_then(|value| value.as_bool())
-      .unwrap_or(false)
-  }
-
-  async fn demo_worker_assets_available() -> bool {
-    fetch_ok("./searchlite-demo-worker.mjs").await
-      && fetch_ok("./searchlite-worker-client.mjs").await
-      && fetch_ok("./pkg/searchlite_wasm.js").await
-  }
-
-  fn js_error_type(err: &JsValue) -> Option<String> {
-    js_sys::Reflect::get(err, &JsValue::from_str("type"))
-      .ok()
-      .and_then(|value| value.as_string())
-  }
-
-  fn skip_if_worker_runtime_unavailable(err: &JsValue) -> bool {
-    matches!(
-      js_error_type(err).as_deref(),
-      Some("worker_error")
-        | Some("worker_spawn_error")
-        | Some("worker_module_import_error")
-        | Some("worker_client_init_error")
-    )
-  }
-
-  async fn new_worker_client_instance() -> Result<JsValue, JsValue> {
-    let module_url = js_sys::eval(
-      "new URL('./searchlite-worker-client.mjs', (self.location && self.location.href) || 'http://localhost/').href",
-    )
-    .ok()
-    .and_then(|value| value.as_string())
-    .unwrap_or_else(|| "./searchlite-worker-client.mjs".to_string());
-    let import_expr = format!(
-      "import({})",
-      serde_json::to_string(&module_url)
-        .unwrap_or_else(|_| "'./searchlite-worker-client.mjs'".into())
-    );
-    let module_promise = js_sys::eval(&import_expr)
-      .map_err(|err| js_error("worker_module_import_error", format!("{err:?}")))?
-      .dyn_into::<js_sys::Promise>()
-      .map_err(|_| {
-        js_error(
-          "worker_module_import_error",
-          "import did not return a Promise",
-        )
-      })?;
-    let module = JsFuture::from(module_promise)
-      .await
-      .map_err(|err| js_error("worker_module_import_error", format!("{err:?}")))?;
-    let ctor = js_sys::Reflect::get(&module, &JsValue::from_str("SearchliteWorkerClient"))?
-      .dyn_into::<js_sys::Function>()
-      .map_err(|_| {
-        js_error(
-          "worker_module_import_error",
-          "SearchliteWorkerClient export missing",
-        )
-      })?;
-    js_sys::Reflect::construct(&ctor, &js_sys::Array::new())
-      .map_err(|err| js_error("worker_client_init_error", format!("{err:?}")))
-  }
-
-  fn call_worker_client_method(
-    client: &JsValue,
-    method: &str,
-    args: &js_sys::Array,
-  ) -> Result<js_sys::Promise, JsValue> {
-    let method_fn = js_sys::Reflect::get(client, &JsValue::from_str(method))?
-      .dyn_into::<js_sys::Function>()
-      .map_err(|_| {
-        js_error(
-          "worker_client_method_error",
-          format!("missing method {method}"),
-        )
-      })?;
-    let value = method_fn.apply(client, args)?;
-    value.dyn_into::<js_sys::Promise>().map_err(|_| {
-      js_error(
-        "worker_client_method_error",
-        format!("method {method} did not return a Promise"),
-      )
-    })
-  }
-
-  fn spawn_demo_worker() -> Result<web_sys::Worker, JsValue> {
-    let worker_js = js_sys::eval("new Worker('./searchlite-demo-worker.mjs', { type: 'module' })")
-      .map_err(|err| js_error("worker_spawn_error", format!("{err:?}")))?;
-    worker_js
-      .dyn_into::<web_sys::Worker>()
-      .map_err(|_| js_error("worker_spawn_error", "failed to cast JS worker"))
-  }
-
-  async fn worker_call(
-    worker: &web_sys::Worker,
-    id: u32,
-    action: &str,
-    payload: JsValue,
-  ) -> Result<JsValue, JsValue> {
-    let (tx, rx) = oneshot::channel::<Result<JsValue, JsValue>>();
-    let tx = Rc::new(RefCell::new(Some(tx)));
-
-    let message_tx = tx.clone();
-    let message_worker = worker.clone();
-    let onmessage =
-      Closure::<dyn FnMut(web_sys::MessageEvent)>::new(move |event: web_sys::MessageEvent| {
-        let data = event.data();
-        let msg_id = js_sys::Reflect::get(&data, &JsValue::from_str("id"))
-          .ok()
-          .and_then(|raw| raw.as_f64())
-          .map(|raw| raw as u32);
-        if msg_id != Some(id) {
-          return;
-        }
-        let ok = js_sys::Reflect::get(&data, &JsValue::from_str("ok"))
-          .ok()
-          .and_then(|raw| raw.as_bool())
-          .unwrap_or(false);
-        let key = if ok { "payload" } else { "error" };
-        let value = js_sys::Reflect::get(&data, &JsValue::from_str(key)).unwrap_or(JsValue::NULL);
-        if let Some(sender) = message_tx.borrow_mut().take() {
-          let _ = sender.send(if ok { Ok(value) } else { Err(value) });
-        }
-        message_worker.set_onmessage(None);
-        message_worker.set_onerror(None);
-      });
-
-    let error_tx = tx.clone();
-    let error_worker = worker.clone();
-    let onerror = Closure::<dyn FnMut(web_sys::Event)>::new(move |event: web_sys::Event| {
-      let reason = js_sys::Reflect::get(event.as_ref(), &JsValue::from_str("message"))
-        .ok()
-        .and_then(|raw| raw.as_string())
-        .unwrap_or_else(|| "worker runtime error".to_string());
-      if let Some(sender) = error_tx.borrow_mut().take() {
-        let _ = sender.send(Err(js_error("worker_error", reason)));
-      }
-      error_worker.set_onmessage(None);
-      error_worker.set_onerror(None);
-    });
-
-    worker.set_onmessage(Some(onmessage.as_ref().unchecked_ref()));
-    worker.set_onerror(Some(onerror.as_ref().unchecked_ref()));
-    onmessage.forget();
-    onerror.forget();
-
-    let msg = js_sys::Object::new();
-    js_sys::Reflect::set(
-      &msg,
-      &JsValue::from_str("id"),
-      &JsValue::from_f64(f64::from(id)),
-    )
-    .unwrap();
-    js_sys::Reflect::set(
-      &msg,
-      &JsValue::from_str("action"),
-      &JsValue::from_str(action),
-    )
-    .unwrap();
-    js_sys::Reflect::set(&msg, &JsValue::from_str("payload"), &payload).unwrap();
-    worker.post_message(&msg)?;
-
-    match rx.await {
-      Ok(result) => result,
-      Err(_) => Err(js_error(
-        "worker_channel_closed",
-        "worker response channel closed",
-      )),
-    }
   }
 
   struct MigrationFailureGuard;
@@ -2869,438 +2679,6 @@ mod tests {
   }
 
   #[wasm_bindgen_test]
-  async fn search_request_roundtrip() {
-    let db = unique_db("searchlite-search-request");
-    let schema = Schema::default_text_body();
-    let schema_json = serde_json::to_string(&schema).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello wasm" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    let request = SearchRequest {
-      query: "hello".into(),
-      fields: None,
-      filter: None,
-      limit: 5,
-      from: 0,
-      return_hits: true,
-      candidate_size: None,
-      #[cfg(feature = "vectors")]
-      max_global_vector_candidates: None,
-      sort: vec![],
-      cursor: None,
-      search_after: None,
-      execution: ExecutionStrategy::Wand,
-      bmw_block_size: None,
-      fuzzy: None,
-      track_total_hits: None,
-      #[cfg(feature = "vectors")]
-      vector_query: None,
-
-      #[cfg(feature = "vectors")]
-      vector_filter: None,
-      return_stored: true,
-      highlight_field: None,
-      highlight: None,
-      collapse: None,
-      aggs: BTreeMap::new(),
-      suggest: BTreeMap::new(),
-      rescore: None,
-      explain: false,
-      profile: false,
-    };
-    let request_json = serde_json::to_string(&request).unwrap();
-    let result = idx.search_request(request_json).unwrap();
-    let result_json: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hits = result_json["hits"].as_array().unwrap();
-    assert_eq!(hits.len(), 1);
-  }
-
-  #[wasm_bindgen_test]
-  async fn delete_document_roundtrip() {
-    let db = unique_db("searchlite-delete-document");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![
-      serde_json::json!({ "_id": "doc-1", "body": "alpha token" }),
-      serde_json::json!({ "_id": "doc-2", "body": "beta token" }),
-    ];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    idx.delete_document("doc-1".to_string()).unwrap();
-    idx.commit().await.unwrap();
-
-    let deleted = idx.search("alpha".to_string(), 5, Some(true)).unwrap();
-    let deleted_json: serde_json::Value = serde_wasm_bindgen::from_value(deleted).unwrap();
-    assert_eq!(deleted_json["hits"].as_array().unwrap().len(), 0);
-
-    let retained = idx.search("beta".to_string(), 5, Some(true)).unwrap();
-    let retained_json: serde_json::Value = serde_wasm_bindgen::from_value(retained).unwrap();
-    assert_eq!(retained_json["hits"].as_array().unwrap().len(), 1);
-  }
-
-  #[wasm_bindgen_test]
-  async fn delete_documents_roundtrip() {
-    let db = unique_db("searchlite-delete-documents");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![
-      serde_json::json!({ "_id": "doc-1", "body": "alpha token" }),
-      serde_json::json!({ "_id": "doc-2", "body": "beta token" }),
-      serde_json::json!({ "_id": "doc-3", "body": "gamma token" }),
-    ];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let ids = vec!["doc-1", "doc-3"];
-    idx
-      .delete_documents(serde_wasm_bindgen::to_value(&ids).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let alpha = idx.search("alpha".to_string(), 5, Some(true)).unwrap();
-    let alpha_json: serde_json::Value = serde_wasm_bindgen::from_value(alpha).unwrap();
-    assert_eq!(alpha_json["hits"].as_array().unwrap().len(), 0);
-
-    let gamma = idx.search("gamma".to_string(), 5, Some(true)).unwrap();
-    let gamma_json: serde_json::Value = serde_wasm_bindgen::from_value(gamma).unwrap();
-    assert_eq!(gamma_json["hits"].as_array().unwrap().len(), 0);
-
-    let beta = idx.search("beta".to_string(), 5, Some(true)).unwrap();
-    let beta_json: serde_json::Value = serde_wasm_bindgen::from_value(beta).unwrap();
-    assert_eq!(beta_json["hits"].as_array().unwrap().len(), 1);
-  }
-
-  #[wasm_bindgen_test]
-  async fn delete_documents_rejects_non_string_ids() {
-    let db = unique_db("searchlite-delete-invalid");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-
-    let invalid = serde_json::json!(["doc-1", 42]);
-    let err = idx
-      .delete_documents(serde_wasm_bindgen::to_value(&invalid).unwrap())
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_doc_id_batch");
-  }
-
-  #[wasm_bindgen_test]
-  async fn delete_document_rejects_invalid_id() {
-    let db = unique_db("searchlite-del-invalid-id");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let err = idx.delete_document("  ".to_string()).unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_id");
-  }
-
-  #[wasm_bindgen_test]
-  async fn delete_documents_rejects_invalid_ids() {
-    let db = unique_db("searchlite-dels-invalid-id");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let ids = vec!["valid-doc", "\n"];
-    let err = idx
-      .delete_documents(serde_wasm_bindgen::to_value(&ids).unwrap())
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_id");
-  }
-
-  #[wasm_bindgen_test]
-  async fn update_document_set_and_unset_roundtrip() {
-    let db = unique_db("searchlite-update-document");
-    let mut schema = Schema::default_text_body();
-    schema
-      .keyword_fields
-      .push(searchlite_core::api::types::KeywordField {
-        name: "category".to_string(),
-        stored: true,
-        indexed: true,
-        fast: true,
-        nullable: true,
-      });
-    let schema_json = serde_json::to_string(&schema).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({
-      "_id": "doc-1",
-      "body": "hello patch",
-      "category": "guide"
-    })];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let patch = serde_json::json!({
-      "id": "doc-1",
-      "set": { "body": "updated patch" },
-      "unset": ["category"]
-    });
-    idx
-      .update_document(serde_wasm_bindgen::to_value(&patch).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let result = idx.search("updated".to_string(), 5, Some(true)).unwrap();
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hits = parsed["hits"].as_array().unwrap();
-    assert_eq!(hits.len(), 1);
-    assert_eq!(hits[0]["fields"]["body"], "updated patch");
-    assert!(hits[0]["fields"]["category"].is_null());
-  }
-
-  #[wasm_bindgen_test]
-  async fn update_document_rejects_missing_patch() {
-    let db = unique_db("searchlite-update-missing-patch");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-
-    let patch = serde_json::json!({ "id": "doc-1" });
-    let err = idx
-      .update_document(serde_wasm_bindgen::to_value(&patch).unwrap())
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "missing_patch");
-  }
-
-  #[wasm_bindgen_test]
-  async fn update_document_rejects_invalid_id() {
-    let db = unique_db("searchlite-update-invalid-id");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-
-    let patch = serde_json::json!({
-      "id": "   ",
-      "set": { "body": "x" }
-    });
-    let err = idx
-      .update_document(serde_wasm_bindgen::to_value(&patch).unwrap())
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_id");
-  }
-
-  #[wasm_bindgen_test]
-  async fn update_document_reports_not_found() {
-    let db = unique_db("searchlite-update-not-found");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-
-    let patch = serde_json::json!({
-      "id": "missing",
-      "set": { "body": "x" }
-    });
-    let err = idx
-      .update_document(serde_wasm_bindgen::to_value(&patch).unwrap())
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "document_not_found");
-  }
-
-  #[wasm_bindgen_test]
-  async fn update_document_rejects_unknown_field() {
-    let db = unique_db("searchlite-update-unknown-field");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({
-      "_id": "doc-1",
-      "body": "hello patch"
-    })];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let patch = serde_json::json!({
-      "id": "doc-1",
-      "set": { "unknown": "x" }
-    });
-    let err = idx
-      .update_document(serde_wasm_bindgen::to_value(&patch).unwrap())
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "update_failed");
-    assert!(payload.reason.contains("unknown field"));
-  }
-
-  #[wasm_bindgen_test]
-  async fn mget_returns_found_missing_and_preserves_order() {
-    let db = unique_db("searchlite-mget-order");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![
-      serde_json::json!({ "_id": "doc-1", "body": "alpha" }),
-      serde_json::json!({ "_id": "doc-2", "body": "beta" }),
-    ];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let req = serde_json::json!({
-      "ids": ["doc-2", "missing", "doc-1"],
-      "return_stored": true
-    });
-    let result = idx
-      .mget(serde_wasm_bindgen::to_value(&req).unwrap())
-      .unwrap();
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let out = parsed["docs"].as_array().unwrap();
-    assert_eq!(out.len(), 3);
-    assert_eq!(out[0]["doc_id"], "doc-2");
-    assert_eq!(out[1]["doc_id"], "missing");
-    assert_eq!(out[2]["doc_id"], "doc-1");
-    assert_eq!(out[0]["found"], true);
-    assert_eq!(out[1]["found"], false);
-    assert_eq!(out[2]["found"], true);
-    assert_eq!(out[0]["_source"]["body"], "beta");
-    assert!(out[1]["_source"].is_null());
-    assert_eq!(out[2]["_source"]["body"], "alpha");
-  }
-
-  #[wasm_bindgen_test]
-  async fn mget_respects_return_stored_false() {
-    let db = unique_db("searchlite-mget-return-stored-false");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "alpha" })];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let req = serde_json::json!({
-      "ids": ["doc-1"],
-      "return_stored": false
-    });
-    let result = idx
-      .mget(serde_wasm_bindgen::to_value(&req).unwrap())
-      .unwrap();
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let out = parsed["docs"].as_array().unwrap();
-    assert_eq!(out.len(), 1);
-    assert_eq!(out[0]["found"], true);
-    assert!(out[0]["_source"].is_null());
-  }
-
-  #[wasm_bindgen_test]
-  async fn mget_rejects_invalid_ids() {
-    let db = unique_db("searchlite-mget-invalid-id");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let req = serde_json::json!({
-      "ids": ["doc-1", "  "],
-      "return_stored": true
-    });
-    let err = idx
-      .mget(serde_wasm_bindgen::to_value(&req).unwrap())
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_id");
-  }
-
-  #[wasm_bindgen_test]
-  async fn multi_search_returns_ordered_results() {
-    let db = unique_db("searchlite-multi-search");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![
-      serde_json::json!({ "_id": "doc-1", "body": "alpha token" }),
-      serde_json::json!({ "_id": "doc-2", "body": "beta token" }),
-    ];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let req = serde_json::json!({
-      "searches": [
-        { "query": "alpha", "limit": 5, "return_stored": true },
-        { "query": "beta", "limit": 5, "return_stored": true }
-      ]
-    });
-    let result = idx
-      .multi_search(serde_wasm_bindgen::to_value(&req).unwrap())
-      .unwrap();
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let results = parsed["results"].as_array().unwrap();
-    assert_eq!(results.len(), 2);
-    assert_eq!(results[0]["hits"][0]["fields"]["body"], "alpha token");
-    assert_eq!(results[1]["hits"][0]["fields"]["body"], "beta token");
-  }
-
-  #[wasm_bindgen_test]
-  async fn multi_search_rejects_invalid_request() {
-    let db = unique_db("searchlite-multi-search-invalid");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-
-    let req = serde_json::json!({ "searches": "not-an-array" });
-    let err = idx
-      .multi_search(serde_wasm_bindgen::to_value(&req).unwrap())
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_multi_search_request");
-  }
-
-  #[wasm_bindgen_test]
-  async fn compact_stats_and_inspect_roundtrip() {
-    let db = unique_db("searchlite-maintenance");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db.clone(), schema_json, None)
-      .await
-      .unwrap();
-
-    let docs_a = vec![serde_json::json!({ "_id": "doc-1", "body": "alpha" })];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs_a).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let docs_b = vec![serde_json::json!({ "_id": "doc-2", "body": "beta" })];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs_b).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let stats_before_js = idx.stats().unwrap();
-    let stats_before: StatsResponse = serde_wasm_bindgen::from_value(stats_before_js).unwrap();
-    assert_eq!(stats_before.documents, 2);
-    assert_eq!(stats_before.deleted_documents, 0);
-    assert_eq!(stats_before.index_name, db);
-    assert!(stats_before.segments >= 2);
-
-    let compact_js = idx.compact().await.unwrap();
-    let compact: CompactResponse = serde_wasm_bindgen::from_value(compact_js).unwrap();
-    assert!(compact.compacted);
-
-    let stats_after_js = idx.stats().unwrap();
-    let stats_after: StatsResponse = serde_wasm_bindgen::from_value(stats_after_js).unwrap();
-    assert_eq!(stats_after.documents, 2);
-    assert_eq!(stats_after.deleted_documents, 0);
-    assert_eq!(stats_after.segments, 1);
-
-    let inspect_js = idx.inspect().unwrap();
-    let inspect: InspectResponse = serde_wasm_bindgen::from_value(inspect_js).unwrap();
-    assert!(inspect.manifest.write_key.is_none());
-    assert_eq!(inspect.manifest.segments.len(), stats_after.segments);
-    assert!(inspect
-      .manifest
-      .segments
-      .iter()
-      .all(|seg| seg.write_binding_b64.is_none()));
-  }
-
-  #[wasm_bindgen_test]
   async fn wasm_core_parity_search_mget_multi_search_update_delete() {
     let db = unique_db("searchlite-parity");
     let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
@@ -3420,530 +2798,6 @@ mod tests {
   }
 
   #[wasm_bindgen_test]
-  async fn search_defaults_skip_stored_fields() {
-    let db = unique_db("searchlite-wasm-return-stored-default");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello wasm" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    let result = idx.search("hello".to_string(), 5, None).unwrap();
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hit = &parsed["hits"][0];
-    assert!(hit["fields"].is_null());
-  }
-
-  #[wasm_bindgen_test]
-  async fn search_request_value_respects_return_stored() {
-    let db = unique_db("searchlite-wasm-return-stored");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello wasm" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-
-    let request = serde_json::json!({
-      "query": "hello",
-      "limit": 5,
-      "return_stored": true
-    });
-    let request_js = serde_wasm_bindgen::to_value(&request).unwrap();
-    let result = idx.search_request_value(request_js).unwrap();
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hit = &parsed["hits"][0];
-    assert!(hit["fields"].is_object());
-  }
-
-  #[wasm_bindgen_test]
-  async fn search_request_value_controlled_rejects_invalid_timeout() {
-    let db = unique_db("searchlite-controlled-invalid-timeout");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let req = serde_json::json!({ "query": "hello", "limit": 5, "return_stored": true });
-    let err = idx
-      .search_request_value_controlled(
-        serde_wasm_bindgen::to_value(&req).unwrap(),
-        None,
-        Some(-1.0),
-      )
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_timeout");
-  }
-
-  #[wasm_bindgen_test]
-  async fn search_request_value_controlled_aborts_with_preaborted_signal() {
-    let db = unique_db("searchlite-controlled-abort");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let req = serde_json::json!({ "query": "hello", "limit": 5, "return_stored": true });
-    let controller = web_sys::AbortController::new().unwrap();
-    controller.abort();
-    let err = idx
-      .search_request_value_controlled(
-        serde_wasm_bindgen::to_value(&req).unwrap(),
-        Some(controller.signal()),
-        None,
-      )
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "aborted");
-  }
-
-  #[wasm_bindgen_test]
-  async fn search_request_value_controlled_times_out() {
-    let db = unique_db("searchlite-controlled-timeout");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let req = serde_json::json!({ "query": "hello", "limit": 5, "return_stored": true });
-    let err = idx
-      .search_request_value_controlled(serde_wasm_bindgen::to_value(&req).unwrap(), None, Some(0.0))
-      .unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "timeout");
-  }
-
-  #[wasm_bindgen_test]
-  async fn search_request_value_async_roundtrip() {
-    let db = unique_db("searchlite-async-search");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello wasm async" })];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-    let req = serde_json::json!({ "query": "hello", "limit": 5, "return_stored": true });
-    let result = idx
-      .search_request_value_async(serde_wasm_bindgen::to_value(&req).unwrap(), None, None)
-      .await
-      .unwrap();
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hits = parsed["hits"].as_array().unwrap();
-    assert_eq!(hits.len(), 1);
-  }
-
-  #[wasm_bindgen_test]
-  async fn worker_search_request_keeps_main_thread_responsive() {
-    if !demo_worker_assets_available().await {
-      return;
-    }
-    let db = unique_db("searchlite-worker-responsive");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let worker = spawn_demo_worker().unwrap();
-
-    let init_payload = serde_json::json!({
-      "dbName": db,
-      "schemaJson": schema_json,
-      "storage": "indexeddb",
-    });
-    let init_result = worker_call(
-      &worker,
-      1,
-      "init_index",
-      serde_wasm_bindgen::to_value(&init_payload).unwrap(),
-    )
-    .await;
-    if let Err(err) = init_result {
-      if skip_if_worker_runtime_unavailable(&err) {
-        worker.terminate();
-        return;
-      }
-      panic!("worker init failed: {err:?}");
-    }
-
-    let docs_payload = serde_json::json!({
-      "docs": [{ "_id": "doc-1", "body": "hello from worker" }],
-    });
-    worker_call(
-      &worker,
-      2,
-      "add_documents",
-      serde_wasm_bindgen::to_value(&docs_payload).unwrap(),
-    )
-    .await
-    .unwrap();
-
-    let timer_fired = Rc::new(Cell::new(false));
-    let timer_fired_ref = Rc::clone(&timer_fired);
-    set_timeout_once(20, move || timer_fired_ref.set(true));
-
-    let request_payload = serde_json::json!({
-      "request": { "query": "hello", "limit": 5, "return_stored": true },
-      "delayMs": 200,
-      "timeoutMs": 800,
-    });
-    let result = match worker_call(
-      &worker,
-      3,
-      "search_request",
-      serde_wasm_bindgen::to_value(&request_payload).unwrap(),
-    )
-    .await
-    {
-      Ok(result) => result,
-      Err(err) if skip_if_worker_runtime_unavailable(&err) => {
-        worker.terminate();
-        let _ = Searchlite::drop_index(db).await;
-        return;
-      }
-      Err(err) => panic!("worker search failed: {err:?}"),
-    };
-    let parsed: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    assert_eq!(parsed["hits"].as_array().unwrap().len(), 1);
-    assert!(
-      timer_fired.get(),
-      "main thread timer should fire while worker search is in-flight"
-    );
-
-    worker.terminate();
-    let db_name = init_payload["dbName"].as_str().unwrap().to_string();
-    Searchlite::drop_index(db_name).await.unwrap();
-  }
-
-  #[wasm_bindgen_test]
-  async fn worker_search_request_timeout_returns_typed_error() {
-    if !demo_worker_assets_available().await {
-      return;
-    }
-    let db = unique_db("searchlite-worker-timeout");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let worker = spawn_demo_worker().unwrap();
-
-    let init_payload = serde_json::json!({
-      "dbName": db,
-      "schemaJson": schema_json,
-      "storage": "indexeddb",
-    });
-    let init_result = worker_call(
-      &worker,
-      1,
-      "init_index",
-      serde_wasm_bindgen::to_value(&init_payload).unwrap(),
-    )
-    .await;
-    if let Err(err) = init_result {
-      if skip_if_worker_runtime_unavailable(&err) {
-        worker.terminate();
-        return;
-      }
-      panic!("worker init failed: {err:?}");
-    }
-
-    let docs_payload = serde_json::json!({
-      "docs": [{ "_id": "doc-1", "body": "timeout path" }],
-    });
-    worker_call(
-      &worker,
-      2,
-      "add_documents",
-      serde_wasm_bindgen::to_value(&docs_payload).unwrap(),
-    )
-    .await
-    .unwrap();
-
-    let request_payload = serde_json::json!({
-      "request": { "query": "timeout", "limit": 5, "return_stored": true },
-      "delayMs": 200,
-      "timeoutMs": 20,
-    });
-    let err = match worker_call(
-      &worker,
-      3,
-      "search_request",
-      serde_wasm_bindgen::to_value(&request_payload).unwrap(),
-    )
-    .await
-    {
-      Ok(_) => panic!("expected timeout error from worker search"),
-      Err(err) if skip_if_worker_runtime_unavailable(&err) => {
-        worker.terminate();
-        let _ = Searchlite::drop_index(db).await;
-        return;
-      }
-      Err(err) => err,
-    };
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "timeout");
-
-    worker.terminate();
-    let db_name = init_payload["dbName"].as_str().unwrap().to_string();
-    Searchlite::drop_index(db_name).await.unwrap();
-  }
-
-  #[wasm_bindgen_test]
-  async fn worker_client_search_request_abort_returns_typed_error() {
-    if !demo_worker_assets_available().await {
-      return;
-    }
-    let db = unique_db("searchlite-worker-client-abort");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let client = match new_worker_client_instance().await {
-      Ok(client) => client,
-      Err(err) => {
-        if skip_if_worker_runtime_unavailable(&err) {
-          return;
-        }
-        panic!("worker client init failed: {err:?}");
-      }
-    };
-
-    let init_args = js_sys::Array::new();
-    init_args.push(&JsValue::from_str(&db));
-    init_args.push(&JsValue::from_str(&schema_json));
-    init_args.push(&JsValue::from_str("indexeddb"));
-    let init_res =
-      JsFuture::from(call_worker_client_method(&client, "initIndex", &init_args).unwrap()).await;
-    if let Err(err) = init_res {
-      if skip_if_worker_runtime_unavailable(&err) {
-        return;
-      }
-      panic!("worker client initIndex failed: {err:?}");
-    }
-
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "abort path" })];
-    let add_args = js_sys::Array::new();
-    add_args.push(&serde_wasm_bindgen::to_value(&docs).unwrap());
-    JsFuture::from(call_worker_client_method(&client, "addDocuments", &add_args).unwrap())
-      .await
-      .unwrap();
-
-    let controller = web_sys::AbortController::new().unwrap();
-    let controller_for_timeout = controller.clone();
-    set_timeout_once(20, move || controller_for_timeout.abort());
-
-    let request = serde_json::json!({ "query": "abort", "limit": 5, "return_stored": true });
-    let options = js_sys::Object::new();
-    js_sys::Reflect::set(
-      &options,
-      &JsValue::from_str("delayMs"),
-      &JsValue::from_f64(200.0),
-    )
-    .unwrap();
-    js_sys::Reflect::set(
-      &options,
-      &JsValue::from_str("timeoutMs"),
-      &JsValue::from_f64(800.0),
-    )
-    .unwrap();
-    js_sys::Reflect::set(
-      &options,
-      &JsValue::from_str("signal"),
-      controller.signal().as_ref(),
-    )
-    .unwrap();
-    let search_args = js_sys::Array::new();
-    search_args.push(&serde_wasm_bindgen::to_value(&request).unwrap());
-    search_args.push(options.as_ref());
-    let err = match JsFuture::from(
-      call_worker_client_method(&client, "searchRequest", &search_args).unwrap(),
-    )
-    .await
-    {
-      Ok(_) => panic!("expected abort error from worker client search"),
-      Err(err) if skip_if_worker_runtime_unavailable(&err) => {
-        return;
-      }
-      Err(err) => err,
-    };
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "aborted");
-
-    let dispose_args = js_sys::Array::new();
-    JsFuture::from(call_worker_client_method(&client, "dispose", &dispose_args).unwrap())
-      .await
-      .unwrap();
-    Searchlite::drop_index(db).await.unwrap();
-  }
-
-  #[wasm_bindgen_test]
-  async fn worker_client_search_request_rejects_invalid_timeout() {
-    if !demo_worker_assets_available().await {
-      return;
-    }
-    let db = unique_db("searchlite-worker-client-invalid-timeout");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let client = match new_worker_client_instance().await {
-      Ok(client) => client,
-      Err(err) => {
-        if skip_if_worker_runtime_unavailable(&err) {
-          return;
-        }
-        panic!("worker client init failed: {err:?}");
-      }
-    };
-
-    let init_args = js_sys::Array::new();
-    init_args.push(&JsValue::from_str(&db));
-    init_args.push(&JsValue::from_str(&schema_json));
-    init_args.push(&JsValue::from_str("indexeddb"));
-    let init_res =
-      JsFuture::from(call_worker_client_method(&client, "initIndex", &init_args).unwrap()).await;
-    if let Err(err) = init_res {
-      if skip_if_worker_runtime_unavailable(&err) {
-        return;
-      }
-      panic!("worker client initIndex failed: {err:?}");
-    }
-
-    let request = serde_json::json!({ "query": "timeout", "limit": 5, "return_stored": true });
-    let options = js_sys::Object::new();
-    js_sys::Reflect::set(
-      &options,
-      &JsValue::from_str("timeoutMs"),
-      &JsValue::from_f64(-1.0),
-    )
-    .unwrap();
-    let search_args = js_sys::Array::new();
-    search_args.push(&serde_wasm_bindgen::to_value(&request).unwrap());
-    search_args.push(options.as_ref());
-    let err = match JsFuture::from(
-      call_worker_client_method(&client, "searchRequest", &search_args).unwrap(),
-    )
-    .await
-    {
-      Ok(_) => panic!("expected invalid_timeout error from worker client search"),
-      Err(err) if skip_if_worker_runtime_unavailable(&err) => {
-        return;
-      }
-      Err(err) => err,
-    };
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "invalid_timeout");
-    assert!(payload.reason.contains("timeoutMs"));
-
-    let dispose_args = js_sys::Array::new();
-    JsFuture::from(call_worker_client_method(&client, "dispose", &dispose_args).unwrap())
-      .await
-      .unwrap();
-    Searchlite::drop_index(db).await.unwrap();
-  }
-
-  #[cfg(not(feature = "threads"))]
-  #[wasm_bindgen_test]
-  async fn init_threads_without_feature_returns_typed_error() {
-    let db = unique_db("searchlite-threads-disabled");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db.clone(), schema_json, None)
-      .await
-      .unwrap();
-    let err = idx.init_threads(None).await.unwrap_err();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "threads_feature_disabled");
-    Searchlite::drop_index(db).await.unwrap();
-  }
-
-  #[wasm_bindgen_test]
-  async fn init_reuses_existing_index() {
-    let db = unique_db("searchlite-reopen");
-    let schema = Schema::default_text_body();
-    let schema_json = serde_json::to_string(&schema).unwrap();
-    let idx = Searchlite::init(db.clone(), schema_json.clone(), None)
-      .await
-      .unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "hello reopen" })];
-    let docs_js = serde_wasm_bindgen::to_value(&docs).unwrap();
-    idx.add_documents(docs_js).unwrap();
-    idx.commit().await.unwrap();
-    drop(idx);
-
-    let reopened = Searchlite::init(db, schema_json, None).await.unwrap();
-    let result = reopened.search("hello".to_string(), 5, None).unwrap();
-    let result_json: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hits = result_json["hits"].as_array().unwrap();
-    assert_eq!(hits.len(), 1);
-  }
-
-  #[wasm_bindgen_test]
-  async fn list_indexes_includes_initialized_db() {
-    let db = unique_db("searchlite-list-indexes");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db.clone(), schema_json, None)
-      .await
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let indexes_js = Searchlite::list_indexes().await.unwrap();
-    let indexes: Vec<IndexRegistryEntry> = serde_wasm_bindgen::from_value(indexes_js).unwrap();
-    assert!(indexes.iter().any(|entry| entry.db_name == db));
-  }
-
-  #[wasm_bindgen_test]
-  async fn clear_index_resets_contents() {
-    let db = unique_db("searchlite-clear-index");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db.clone(), schema_json.clone(), None)
-      .await
-      .unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "clear me" })];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    Searchlite::clear_index(db.clone()).await.unwrap();
-
-    // Verify the index remains discoverable in the registry after clear.
-    let indexes_js = Searchlite::list_indexes().await.unwrap();
-    let indexes: Vec<IndexRegistryEntry> = serde_wasm_bindgen::from_value(indexes_js).unwrap();
-    assert!(
-      indexes.iter().any(|entry| entry.db_name == db),
-      "clear_index should preserve the registry entry"
-    );
-
-    let reopened = Searchlite::init(db, schema_json, None).await.unwrap();
-    let result = reopened.search("clear".to_string(), 5, None).unwrap();
-    let result_json: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hits = result_json["hits"].as_array().unwrap();
-    assert_eq!(hits.len(), 0);
-  }
-
-  #[wasm_bindgen_test]
-  async fn drop_index_removes_registry_entry() {
-    let db = unique_db("searchlite-drop-index");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let idx = Searchlite::init(db.clone(), schema_json.clone(), None)
-      .await
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    Searchlite::drop_index(db.clone()).await.unwrap();
-
-    let indexes_js = Searchlite::list_indexes().await.unwrap();
-    let indexes: Vec<IndexRegistryEntry> = serde_wasm_bindgen::from_value(indexes_js).unwrap();
-    assert!(!indexes.iter().any(|entry| entry.db_name == db));
-
-    // Can recreate after deletion.
-    let recreated = Searchlite::init(db, schema_json, None).await.unwrap();
-    recreated.commit().await.unwrap();
-  }
-
-  #[wasm_bindgen_test]
-  async fn init_rejects_reserved_registry_name() {
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-    let result = Searchlite::init(REGISTRY_DB_NAME.to_string(), schema_json, None).await;
-    assert!(result.is_err(), "expected reserved_name error");
-    let err = result.err().unwrap();
-    let payload: WasmErrorPayload = serde_wasm_bindgen::from_value(err).unwrap();
-    assert_eq!(payload.error_type, "reserved_name");
-  }
-
-  #[wasm_bindgen_test]
-  async fn storage_usage_returns_supported_or_note() {
-    let usage_js = Searchlite::storage_usage().await.unwrap();
-    let usage: StorageUsageResponse = serde_wasm_bindgen::from_value(usage_js).unwrap();
-    assert!(usage.supported || usage.note.is_some());
-    if usage.supported {
-      assert!(
-        usage.usage_bytes.is_some() || usage.quota_bytes.is_some() || usage.persisted.is_some()
-      );
-    }
-  }
-
-  #[wasm_bindgen_test]
   async fn commit_surfaces_quota_exceeded_error_type() {
     let db = unique_db("searchlite-quota-error");
     let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
@@ -3963,40 +2817,6 @@ mod tests {
     assert!(payload.reason.contains("compact()"));
     assert!(payload.reason.contains("cleanup_indexes"));
     Searchlite::drop_index(db).await.unwrap();
-  }
-
-  #[wasm_bindgen_test]
-  async fn cleanup_indexes_drops_only_stale_entries() {
-    let fresh = unique_db("searchlite-cleanup-fresh");
-    let stale = unique_db("searchlite-cleanup-stale");
-    let schema_json = serde_json::to_string(&Schema::default_text_body()).unwrap();
-
-    let fresh_idx = Searchlite::init(fresh.clone(), schema_json.clone(), None)
-      .await
-      .unwrap();
-    fresh_idx.commit().await.unwrap();
-    let stale_idx = Searchlite::init(stale.clone(), schema_json, None)
-      .await
-      .unwrap();
-    stale_idx.commit().await.unwrap();
-
-    let mut stale_entry = get_registry_entry(&stale).await.unwrap().unwrap();
-    stale_entry.updated_at_ms -= 60_000.0;
-    upsert_registry_entry(&stale_entry).await.unwrap();
-
-    let cleanup_js = Searchlite::cleanup_indexes(30_000.0, Some(false))
-      .await
-      .unwrap();
-    let cleanup: CleanupIndexesResponse = serde_wasm_bindgen::from_value(cleanup_js).unwrap();
-    assert!(cleanup.dropped.iter().any(|name| name == &stale));
-    assert!(cleanup.kept.iter().any(|name| name == &fresh));
-
-    let indexes_js = Searchlite::list_indexes().await.unwrap();
-    let indexes: Vec<IndexRegistryEntry> = serde_wasm_bindgen::from_value(indexes_js).unwrap();
-    assert!(indexes.iter().any(|entry| entry.db_name == fresh));
-    assert!(!indexes.iter().any(|entry| entry.db_name == stale));
-
-    Searchlite::drop_index(fresh).await.unwrap();
   }
 
   #[wasm_bindgen_test]
@@ -4032,104 +2852,6 @@ mod tests {
     let hits = result_json["hits"].as_array().unwrap();
     assert_eq!(hits.len(), 1);
     Searchlite::drop_index(db).await.unwrap();
-  }
-
-  #[wasm_bindgen_test]
-  async fn plan_migration_reports_compatibility_and_rebuild() {
-    let db = unique_db("searchlite-plan-migration");
-    let schema_v1 = Schema::default_text_body();
-    let schema_v1_json = serde_json::to_string(&schema_v1).unwrap();
-    let idx = Searchlite::init(db.clone(), schema_v1_json.clone(), None)
-      .await
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let compatible_js = Searchlite::plan_migration(db.clone(), schema_v1_json)
-      .await
-      .unwrap();
-    let compatible: MigrationPlan = serde_wasm_bindgen::from_value(compatible_js).unwrap();
-    assert_eq!(compatible.status, "compatible");
-    assert!(!compatible.rebuild_required);
-
-    let mut schema_v2 = Schema::default_text_body();
-    schema_v2
-      .keyword_fields
-      .push(searchlite_core::api::types::KeywordField {
-        name: "category".to_string(),
-        stored: true,
-        indexed: true,
-        fast: true,
-        nullable: false,
-      });
-    let schema_v2_json = serde_json::to_string(&schema_v2).unwrap();
-    let rebuild_js = Searchlite::plan_migration(db, schema_v2_json)
-      .await
-      .unwrap();
-    let rebuild: MigrationPlan = serde_wasm_bindgen::from_value(rebuild_js).unwrap();
-    assert_eq!(rebuild.status, "rebuild_required");
-    assert!(rebuild.rebuild_required);
-  }
-
-  #[wasm_bindgen_test]
-  async fn migrate_index_creates_missing_index() {
-    let db = unique_db("searchlite-migrate-create");
-    let schema = Schema::default_text_body();
-    let schema_json = serde_json::to_string(&schema).unwrap();
-
-    let created_js = Searchlite::migrate_index(db.clone(), schema_json.clone())
-      .await
-      .unwrap();
-    let created: MigrationExecutionResult = serde_wasm_bindgen::from_value(created_js).unwrap();
-    assert_eq!(created.status, "created");
-    assert!(!created.rebuild_performed);
-
-    let idx = Searchlite::init(db, schema_json, None).await.unwrap();
-    let result = idx.search("anything".to_string(), 5, None).unwrap();
-    let result_json: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hits = result_json["hits"].as_array().unwrap();
-    assert_eq!(hits.len(), 0);
-  }
-
-  #[wasm_bindgen_test]
-  async fn migrate_index_rebuilds_on_schema_change() {
-    let db = unique_db("searchlite-migrate-rebuild");
-    let schema_v1 = Schema::default_text_body();
-    let schema_v1_json = serde_json::to_string(&schema_v1).unwrap();
-    let idx = Searchlite::init(db.clone(), schema_v1_json, None)
-      .await
-      .unwrap();
-    let docs = vec![serde_json::json!({ "_id": "doc-1", "body": "keep me if rollback" })];
-    idx
-      .add_documents(serde_wasm_bindgen::to_value(&docs).unwrap())
-      .unwrap();
-    idx.commit().await.unwrap();
-
-    let mut schema_v2 = Schema::default_text_body();
-    schema_v2
-      .keyword_fields
-      .push(searchlite_core::api::types::KeywordField {
-        name: "category".to_string(),
-        stored: true,
-        indexed: true,
-        fast: true,
-        nullable: false,
-      });
-    let schema_v2_json = serde_json::to_string(&schema_v2).unwrap();
-
-    let rebuild_js = Searchlite::migrate_index(db.clone(), schema_v2_json.clone())
-      .await
-      .unwrap();
-    let rebuild: MigrationExecutionResult = serde_wasm_bindgen::from_value(rebuild_js).unwrap();
-    assert_eq!(rebuild.status, "rebuilt");
-    assert!(rebuild.rebuild_performed);
-
-    let reopened = Searchlite::init(db, schema_v2_json, None).await.unwrap();
-    let result = reopened
-      .search("rollback".to_string(), 5, Some(true))
-      .unwrap();
-    let result_json: serde_json::Value = serde_wasm_bindgen::from_value(result).unwrap();
-    let hits = result_json["hits"].as_array().unwrap();
-    assert_eq!(hits.len(), 0);
   }
 
   #[wasm_bindgen_test]
